+++ src/main/java/com/drzewo97/ballotbox/core/model/candidate/CandidateRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CandidateRepository extends CrudRepository<Candidate, Long> {
+public interface CandidateRepository extends CrudRepository<Candidate, Integer> {
     Boolean existsByName(String name);
     Boolean existsByPollAndUser(Poll poll, User user);
 }
