+++ src/main/java/com/drzewo97/ballotbox/core/model/committeecandidateorder/CommitteeCandidateOrderRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommitteeCandidateOrderRepository extends CrudRepository<CommitteeCandidateOrder, Long> {}
+public interface CommitteeCandidateOrderRepository extends CrudRepository<CommitteeCandidateOrder, Integer> {}
