+++ src/main/java/com/drzewo97/ballotbox/core/model/district/DistrictRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DistrictRepository extends CrudRepository<District, Long> {
+public interface DistrictRepository extends CrudRepository<District, Integer> {
 	Boolean existsByNameAndCountry_Name(String name, String countryName);
 }
