+++ src/main/java/com/drzewo97/ballotbox/core/service/ward/WardService.java
@@ -1,5 +1,5 @@
 package com.drzewo97.ballotbox.core.service.ward;
 
 public interface WardService {
-	Boolean isWardAdmin(String username, Long wardId);
+	Boolean isWardAdmin(String username, Integer wardId);
 }
